Move orderbook package outside of tools
