+++ exp/orderbook/edges.go
@@ -1,6 +1,6 @@
 //lint:file-ignore U1000 this package is currently unused but it will be used in a future PR
 
-package main
+package orderbook
 
 import (
 	"math/big"
