+++ exp/orderbook/graph_test.go
@@ -1,4 +1,4 @@
-package main
+package orderbook
 
 import (
 	"bytes"
